+++ test/unit/specs/components/staking/DelegationModal.spec.js
@@ -50,7 +50,7 @@ describe(`DelegationModal`, () => {
     store.state.delegates.delegates = lcdClientMock.state.candidates
     store.commit(`setStakingParameters`, stakingParameters.parameters)
 
-    wrapper.vm.$refs.actionModal.submit = jest.fn(cb => cb())
+    wrapper.vm.$refs.actionModal.open()
   })
 
   describe(`component matches snapshot`, () => {
@@ -59,39 +59,28 @@ describe(`DelegationModal`, () => {
     })
   })
 
-  describe(`only submits on correct form`, () => {
-    describe(`does not submit`, () => {
+  it(`opens`, () => {
+    wrapper.vm.$refs.actionModal.open = jest.fn()
+    wrapper.vm.open()
+    expect(wrapper.vm.$refs.actionModal.open).toHaveBeenCalled()
+  })
+
+  describe(`validation`, () => {
+    describe(`fails`, () => {
       it(`with default values`, () => {
-        wrapper.vm.submitForm = jest.fn()
-        wrapper.vm.validateForm()
-        expect(wrapper.vm.submitForm).not.toHaveBeenCalled()
+        expect(wrapper.vm.validateForm()).toBe(false)
       })
 
-      it(`if the user manually inputs a number greater than the balance`, async () => {
-        wrapper.setData({ amount: 142, password: `1234567890` })
-        wrapper.vm.submitForm = jest.fn()
-        wrapper.vm.validateForm()
-        expect(wrapper.vm.submitForm).not.toHaveBeenCalled()
-
-        await wrapper.vm.$nextTick()
-        let errorMessage = wrapper.find(`input#amount + div`)
-        expect(errorMessage.classes()).toContain(`tm-form-msg--error`)
-      })
-
-      it(`if the password field is empty`, () => {
-        wrapper.setData({ amount: 10, password: `` })
-        wrapper.vm.submitForm = jest.fn()
-        wrapper.vm.validateForm()
-        expect(wrapper.vm.submitForm).not.toHaveBeenCalled()
+      it(`if the user manually inputs a number greater than the balance`, () => {
+        wrapper.setData({ amount: 142 })
+        expect(wrapper.vm.validateForm()).toBe(false)
       })
     })
 
-    describe(`submits`, () => {
+    describe(`succeeds`, () => {
       it(`if the amount is positive and the user has enough balance`, () => {
-        wrapper.setData({ amount: 50, password: `1234567890` })
-        wrapper.vm.submitForm = jest.fn()
-        wrapper.vm.validateForm()
-        expect(wrapper.vm.submitForm).toHaveBeenCalled()
+        wrapper.setData({ amount: 50 })
+        expect(wrapper.vm.validateForm()).toBe(true)
       })
     })
   })
@@ -101,8 +90,8 @@ describe(`DelegationModal`, () => {
       wrapper.vm.$store.dispatch = jest.fn()
       wrapper.vm.$store.commit = jest.fn()
 
-      wrapper.setData({ amount: 50, password: `1234567890` })
-      await wrapper.vm.submitForm()
+      wrapper.setData({ amount: 50 })
+      await wrapper.vm.submitForm(`local`, `1234567890`)
 
       expect(wrapper.vm.$store.dispatch.mock.calls).toEqual([
         [
@@ -112,7 +101,8 @@ describe(`DelegationModal`, () => {
             // validatorSrc: lcdClientMock.state.candidates[1],
             // validatorDst: lcdClientMock.state.candidates[0],
             validator_addr: `cosmosvaladdr15ky9du8a2wlstz6fpx3p4mqpjyrm5ctqzh8yqw`,
-            password: `1234567890`
+            password: `1234567890`,
+            submitType: `local`
           }
         ]
       ])
@@ -132,8 +122,8 @@ describe(`DelegationModal`, () => {
       wrapper.vm.$store.dispatch = jest.fn()
       wrapper.vm.$store.commit = jest.fn()
 
-      wrapper.setData({ amount: 50, password: `1234567890`, selectedIndex: 1 })
-      await wrapper.vm.submitForm()
+      wrapper.setData({ amount: 50, selectedIndex: 1 })
+      await wrapper.vm.submitForm(`local`, `1234567890`)
 
       expect(wrapper.vm.$store.dispatch.mock.calls).toEqual([
         [
@@ -142,6 +132,7 @@ describe(`DelegationModal`, () => {
             amount: `50`,
             validatorSrc: lcdClientMock.state.candidates[0],
             validatorDst: lcdClientMock.state.candidates[0],
+            submitType: `local`,
             password: `1234567890`
           }
         ]
@@ -155,18 +146,6 @@ describe(`DelegationModal`, () => {
             title: `Successful redelegation!`
           }
         ]
-      ])
-    })
-  })
-
-  describe(`closes modal correctly`, () => {
-    it(`X button emits close signal`, () => {
-      wrapper.vm.close()
-      expect(wrapper.emittedByOrder()).toEqual([
-        {
-          name: `update:showDelegationModal`,
-          args: [false]
-        }
       ])
     })
   })